% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavioral_tests.R
\docType{data}
\name{behavioral_tests}
\alias{behavioral_tests}
\alias{behavioral_tests_with_replicates}
\alias{behavioral_tests_no_replicates_median}
\alias{behavioral_tests_no_replicates}
\alias{behavioral_tests_multiblock}
\title{Behavioral tests}
\format{
An object of class \code{data.frame} with 1290 rows and 9 columns.

An object of class \code{data.frame} with 430 rows and 8 columns.

An object of class \code{data.frame} with 430 rows and 8 columns.

An object of class \code{list} of length 5.
}
\usage{
data(behavioral_tests_with_replicates)

data(behavioral_tests_no_replicates_median)

data(behavioral_tests_no_replicates)

data(behavioral_tests_multiblock)
}
\description{
Behavioral tests of mice with immune cell transplantation from multiple
sclerosis patients.
}
\keyword{datasets}
