% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perrot-package.R
\docType{package}
\name{perrot-package}
\alias{perrot}
\alias{perrot-package}
\title{perrot: MS Immunology}
\description{
Behavioral tests and neuronal electrophysiology analysis (linear mixed models) of mice with immune cell transplantation from multiple sclerosis patients.
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

Authors:
\itemize{
  \item Baptiste Criniere-Boizet
}

}
\keyword{internal}
